+++ src/components/list_items_modal/list_item_component.jsx
@@ -4,7 +4,7 @@ import { connect } from "react-redux";
 
 class ListItemComponent extends Component {
     render() {
-        const { citationIndex, sectionId, listItem, dispatch, selectedListItems} = this.props;
+        const { listItemIndex, sectionId, listItem, dispatch, selectedListItems} = this.props;
         let contributors = "";
         let citationLine = "";
         let title = "";
@@ -17,21 +17,21 @@ class ListItemComponent extends Component {
             contributors = listItem.author.join(", ");
             citationLine = listItem.citation_line;
         }
-        
+
         return (
             <li>
                 <div className="checkboxWrap mll">
-                    <label htmlFor={`section_${sectionId}_item_${citationIndex}`}>
+                    <label htmlFor={`section_${sectionId}_item_${listItemIndex}`}>
                         <input
                           className="mylists-single-option"
-                          id={`section_${sectionId}_item_${citationIndex}`}
+                          id={`section_${sectionId}_item_${listItemIndex}`}
                           type="checkbox"
                           value={title}
                           onChange={event => {
                               if (event.target.checked) {
-                                  dispatch(addListItem(sectionId, listItem.formattedCitation, citationIndex));
+                                  dispatch(addListItem(sectionId, listItem.formattedCitation, listItemIndex));
                               } else {
-                                  dispatch(removeListItem(sectionId, citationIndex));
+                                  dispatch(removeListItem(sectionId, listItemIndex));
                               }
                           }} />
                         <span className="pls title lookslikeh2">{title}</span>
@@ -46,7 +46,7 @@ class ListItemComponent extends Component {
 
 ListItemComponent.propTypes = {
     listItem: PropTypes.object.isRequired,
-    citationIndex: PropTypes.number.isRequired,
+    listItemIndex: PropTypes.number.isRequired,
     sectionId: PropTypes.number.isRequired,
     dispatch: PropTypes.func.isRequired
 };
